Питание, Активность, Чат: 21.05.2025

+++ app/src/main/java/com/example/supporthealth/chat/domain/models/ProductAi.kt
@@ -7,5 +7,5 @@ data class ProductAi(
     val protein: Float,
     val fat: Float,
     val carbs: Float,
-    var grams: Int
+    var grams: Float
 )
